feat: add nix flake
