+++ .gitignore
@@ -1,0 +1,6 @@
+# Python artifacts
+/*.egg-info/
+/__pycache__/
+
+# Nix artifacts
+result
